refactor(db): split document repository into cohesive sub-modules
The document repository was a single 3073-line file flagged as a churn
hotspot. Split the one giant `impl DocumentRepository` block into a
`document/` module directory with cohesive sub-modules — folders,
documents, shares, versions, intelligence — plus an `internal` module
for the private row-mapping / token / password helpers.

This is a pure, behavior-preserving move: no SQL, signatures, or logic
changed. Method counts are identical (90 `pub async fn` + `new` + 4
private helpers) and the public path
(`repositories::document::DocumentRepository`) is unchanged. The struct's
`pool` field is now `pub(super)` so the sibling sub-modules can reach it,
and the shared private helpers are `pub(super)` for the same reason.

Each sub-module keeps the file's existing section grouping (the RLS-aware
+ legacy banner blocks), so future edits land in a small focused file
instead of the 3000-line monolith — shrinking the churn surface.
